+++ CropperSample/build.gradle
@@ -1,17 +1,18 @@
-apply plugin: 'android'
+apply plugin: 'com.android.application'
 
 dependencies {
     compile project(':cropper')
+    compile 'com.android.support:appcompat-v7:23.1.0'
 }
 
 android {
 
-    compileSdkVersion 22
-    buildToolsVersion '22.0.1'
+    compileSdkVersion 23
+    buildToolsVersion '23.0.2'
 
     defaultConfig {
         minSdkVersion 7
-        targetSdkVersion 22
+        targetSdkVersion 23
     }
 
     sourceSets {
