--- CropperSample/res/values-normal/dimens.xml
@@ -1,8 +1,0 @@
-<resources>
-
-    <!--
-         Customize dimensions originally defined in res/values/dimens.xml (such as
-         screen margins) for sw600dp devices (e.g. 7" tablets) here.
-    -->
-    <dimen name="textSize">16sp</dimen>
-</resources>
