+++ CropperSample/res/values/dimens.xml
@@ -1,7 +1,5 @@
 <resources>
 
-    <!-- Default screen margins, per the Android Design guidelines. -->
-    <dimen name="activity_horizontal_margin">16dp</dimen>
-    <dimen name="activity_vertical_margin">16dp</dimen>
+    <dimen name="content_padding">16dp</dimen>
 
 </resources>
