+++ CropperSample/res/values/strings.xml
@@ -4,14 +4,13 @@
     <string name="app_name">Cropper Sample</string>
     <string name="title">Cropper</string>
     <string name="action_settings">Settings</string>
-    <string name="fixedAspectRatio">fixedAspectRatio =</string>
+    <string name="fixedAspectRatio">fixedAspectRatio = </string>
     <string name="aspectRatioX">\u002010</string>
-    <string name="aspectRatioXHeader">aspectRatioX =</string>
+    <string name="aspectRatioXHeader">aspectRatioX = </string>
     <string name="aspectRatioY">\u002010</string>
-    <string name="aspectRatioYHeader">aspectRatioY =</string>
+    <string name="aspectRatioYHeader">aspectRatioY = </string>
     <string name="showGuidelines">showGuidelines = </string>
     <string name="crop">Crop</string>
-    <string name="rotate">Rotate</string>
     <string name="croppedImageDesc">The cropped image.</string>
 
     <string-array name="showGuidelinesArray">
