--- CropperSample/res/values/styles.xml
@@ -1,30 +1,0 @@
-<resources>
-
-    <!--
-        Base application theme, dependent on API level. This theme is replaced
-        by AppBaseTheme from res/values-vXX/styles.xml on newer devices.
-    -->
-    <style name="AppBaseTheme" parent="android:Theme.Light">
-        <!--
-            Theme customizations available in newer API levels can go in
-            res/values-vXX/styles.xml, while customizations related to
-            backward-compatibility can go here.
-        -->
-    </style>
-
-    <!-- Application theme. -->
-    <style name="AppTheme" parent="android:Theme.Holo" />
-    <!-- All customizations that are NOT specific to a particular API-level can go here. -->
-
-
-    <!-- Robo Theme. -->
-    <style name="RoboTheme" parent="@android:style/TextAppearance.DeviceDefault.Medium">
-        <item name="android:layout_width">fill_parent</item>
-        <item name="android:layout_height">30dp</item>
-        <item name="android:textColor">#FFFFFF</item>
-        <item name="android:textSize">@dimen/textSize</item>
-        <item name="android:gravity">center</item>
-        <item name="android:layout_marginTop">10dp</item>
-    </style>
-
-</resources>
