+++ CropperSample/src/com/example/croppersample/MainActivity.java
@@ -3,17 +3,13 @@ package com.example.croppersample;
 
 import android.app.Activity;
 import android.graphics.Bitmap;
-import android.graphics.Typeface;
 import android.os.Bundle;
-import android.view.Menu;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.Window;
 import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.CompoundButton;
 import android.widget.CompoundButton.OnCheckedChangeListener;
-import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.SeekBar;
 import android.widget.SeekBar.OnSeekBarChangeListener;
@@ -26,69 +22,31 @@ import com.edmodo.cropper.CropImageView;
 public class MainActivity extends Activity {
 
     // Static final constants
-    private static final int DEFAULT_ASPECT_RATIO_VALUES = 10;
-    private static final int ROTATE_NINETY_DEGREES = 90;
-    private static final String ASPECT_RATIO_X = "ASPECT_RATIO_X";
-    private static final String ASPECT_RATIO_Y = "ASPECT_RATIO_Y";
+    private static final int DEFAULT_ASPECT_RATIO_VALUES = 1;
     private static final int ON_TOUCH = 1;
 
-    // Instance variables
-    private int mAspectRatioX = DEFAULT_ASPECT_RATIO_VALUES;
-    private int mAspectRatioY = DEFAULT_ASPECT_RATIO_VALUES;
-
     Bitmap croppedImage;
-
-    // Saves the state upon rotating the screen/restarting the activity
-    @Override
-    protected void onSaveInstanceState(Bundle bundle) {
-        super.onSaveInstanceState(bundle);
-        bundle.putInt(ASPECT_RATIO_X, mAspectRatioX);
-        bundle.putInt(ASPECT_RATIO_Y, mAspectRatioY);
-    }
-
-    // Restores the state upon rotating the screen/restarting the activity
-    @Override
-    protected void onRestoreInstanceState(Bundle bundle) {
-        super.onRestoreInstanceState(bundle);
-        mAspectRatioX = bundle.getInt(ASPECT_RATIO_X);
-        mAspectRatioY = bundle.getInt(ASPECT_RATIO_Y);
-    }
 
     public void onCreate(Bundle savedInstanceState) {
 
         super.onCreate(savedInstanceState);
-        this.requestWindowFeature(Window.FEATURE_NO_TITLE);
+        requestWindowFeature(Window.FEATURE_NO_TITLE);
         setContentView(R.layout.activity_main);
-
-        // Sets fonts for all
-        Typeface mFont = Typeface.createFromAsset(getAssets(), "Roboto-Thin.ttf");
-        ViewGroup root = (ViewGroup) findViewById(R.id.mylayout);
-        setFont(root, mFont);
 
         // Initialize components of the app
         final CropImageView cropImageView = (CropImageView) findViewById(R.id.CropImageView);
         final SeekBar aspectRatioXSeek = (SeekBar) findViewById(R.id.aspectRatioXSeek);
         final SeekBar aspectRatioYSeek = (SeekBar) findViewById(R.id.aspectRatioYSeek);
         final ToggleButton fixedAspectRatioToggle = (ToggleButton) findViewById(R.id.fixedAspectRatioToggle);
-        Spinner showGuidelinesSpin = (Spinner) findViewById(R.id.showGuidelinesSpin);
-        
+        final Spinner showGuidelinesSpin = (Spinner) findViewById(R.id.showGuidelinesSpin);
+
         // Sets sliders to be disabled until fixedAspectRatio is set
         aspectRatioXSeek.setEnabled(false);
         aspectRatioYSeek.setEnabled(false);
 
         // Set initial spinner value
         showGuidelinesSpin.setSelection(ON_TOUCH);
-        
-        //Sets the rotate button
-        final Button rotateButton = (Button) findViewById(R.id.Button_rotate);
-        rotateButton.setOnClickListener(new View.OnClickListener() {
 
-            @Override
-            public void onClick(View v) {
-                cropImageView.rotateImage(ROTATE_NINETY_DEGREES);
-            }
-        });
-        
         // Sets fixedAspectRatio
         fixedAspectRatioToggle.setOnCheckedChangeListener(new OnCheckedChangeListener() {
 
@@ -98,8 +56,7 @@ public class MainActivity extends Activity {
                 if (isChecked) {
                     aspectRatioXSeek.setEnabled(true);
                     aspectRatioYSeek.setEnabled(true);
-                }
-                else {
+                } else {
                     aspectRatioXSeek.setEnabled(false);
                     aspectRatioYSeek.setEnabled(false);
                 }
@@ -111,16 +68,13 @@ public class MainActivity extends Activity {
 
         // Sets aspectRatioX
         final TextView aspectRatioX = (TextView) findViewById(R.id.aspectRatioX);
+        final TextView aspectRatioY = (TextView) findViewById(R.id.aspectRatioY);
 
         aspectRatioXSeek.setOnSeekBarChangeListener(new OnSeekBarChangeListener() {
             @Override
             public void onProgressChanged(SeekBar aspectRatioXSeek, int progress, boolean fromUser) {
-                try {
-                    mAspectRatioX = progress;
-                    cropImageView.setAspectRatio(progress, mAspectRatioY);
-                    aspectRatioX.setText(" " + progress);
-                } catch (IllegalArgumentException e) {
-                }
+                cropImageView.setAspectRatio(progress, aspectRatioYSeek.getProgress());
+                aspectRatioX.setText(String.valueOf(progress));
             }
 
             @Override
@@ -133,25 +87,22 @@ public class MainActivity extends Activity {
         });
 
         // Sets aspectRatioY
-        final TextView aspectRatioY = (TextView) findViewById(R.id.aspectRatioY);
 
         aspectRatioYSeek.setOnSeekBarChangeListener(new OnSeekBarChangeListener() {
             @Override
             public void onProgressChanged(SeekBar aspectRatioYSeek, int progress, boolean fromUser) {
-                try {
-                    mAspectRatioY = progress;
-                    cropImageView.setAspectRatio(mAspectRatioX, progress);
-                    aspectRatioY.setText(" " + progress);
-                } catch (IllegalArgumentException e) {
-                }
+                cropImageView.setAspectRatio(aspectRatioXSeek.getProgress(), progress);
+                aspectRatioY.setText(String.valueOf(progress));
             }
 
             @Override
             public void onStartTrackingTouch(SeekBar seekBar) {
+                // Do nothing.
             }
 
             @Override
             public void onStopTrackingTouch(SeekBar seekBar) {
+                // Do nothing.
             }
         });
 
@@ -163,7 +114,7 @@ public class MainActivity extends Activity {
             }
 
             public void onNothingSelected(AdapterView<?> adapterView) {
-                return;
+                // Do nothing.
             }
         });
 
@@ -172,36 +123,11 @@ public class MainActivity extends Activity {
 
             @Override
             public void onClick(View v) {
-                croppedImage = cropImageView.getCroppedImage();
-                ImageView croppedImageView = (ImageView) findViewById(R.id.croppedImageView);
+                final Bitmap croppedImage = cropImageView.getCroppedImage();
+                final ImageView croppedImageView = (ImageView) findViewById(R.id.croppedImageView);
                 croppedImageView.setImageBitmap(croppedImage);
             }
         });
 
     }
-
-    /*
-     * Sets the font on all TextViews in the ViewGroup. Searches recursively for
-     * all inner ViewGroups as well. Just add a check for any other views you
-     * want to set as well (EditText, etc.)
-     */
-    public void setFont(ViewGroup group, Typeface font) {
-        int count = group.getChildCount();
-        View v;
-        for (int i = 0; i < count; i++) {
-            v = group.getChildAt(i);
-            if (v instanceof TextView || v instanceof EditText || v instanceof Button) {
-                ((TextView) v).setTypeface(font);
-            } else if (v instanceof ViewGroup)
-                setFont((ViewGroup) v, font);
-        }
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.main, menu);
-        return true;
-    }
-
 }
