+++ cropper/AndroidManifest.xml
@@ -1,6 +1,5 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+<manifest
     package="com.edmodo.cropper"
-    android:versionCode="1"
-    android:versionName="1.0" >
-
-</manifest>
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:versionCode="2"
+    android:versionName="2.0"/>
