+++ cropper/build.gradle
@@ -7,12 +7,12 @@ group = "com.edmodo"
 
 android {
 
-    compileSdkVersion 22
-    buildToolsVersion '22.0.1'
+    compileSdkVersion 23
+    buildToolsVersion '23.0.2'
 
     defaultConfig {
         minSdkVersion 7
-        targetSdkVersion 22
+        targetSdkVersion 23
     }
 
     sourceSets {
