--- cropper/res/layout/crop_image_view.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<merge
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <ImageView
-        android:id="@+id/ImageView_image"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:adjustViewBounds="true"
-        android:scaleType="centerInside"
-        tools:ignore="contentDescription"/>
-
-    <com.edmodo.cropper.cropwindow.CropOverlayView
-        android:id="@+id/CropOverlayView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-
-</merge>
