+++ cropper/res/values/attrs.xml
@@ -1,15 +1,17 @@
 <resources>
 
     <declare-styleable name="CropImageView">
+
         <attr name="guidelines">
             <enum name="on" value="2"/>
             <enum name="onTouch" value="1"/>
             <enum name="off" value="0"/>
         </attr>
+
         <attr name="fixAspectRatio" format="boolean"/>
         <attr name="aspectRatioX" format="integer"/>
-        <attr name="aspectRatioY" format= "integer"/>
-        <attr name="imageResource" format="reference"/>
+        <attr name="aspectRatioY" format="integer"/>
+
     </declare-styleable>
 
 </resources>
