--- cropper/res/values/strings.xml
@@ -1,5 +1,0 @@
-<resources>
-
-    <string name="app_name">Cropper</string>
-
-</resources>
